Rename feature from chat to assist in the codebase
