+++ bundles/org.openhab.ui/web/src/assets/i18n/assist/en.json
