+++ bundles/org.openhab.ui/web/src/components/assist/assist-tool-return.vue
