+++ bundles/org.openhab.ui/web/src/components/assist/assist-welcome.vue
